pull functionality into current body_modifier
