--- body/test.json
@@ -1,6 +1,0 @@
-{
-  "widget": {
-    "foo": "on",
-    "bar": "off"
-  }
-}
